# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type Category = (typeof Category)[keyof typeof Category];

export const Category = {
  top: "top",
  bottom: "bottom",
  shoes: "shoes",
  outerwear: "outerwear",
  accessory: "accessory",
} as const;

export interface ClothingItem {
  id: number;
  name: string;
  category: Category;
  color?: string | null;
  imagePath: string;
  createdAt: string;
}

export interface CreateItemBody {
  /** @minLength 1 */
  name: string;
  category: Category;
  color?: string | null;
  /** @minLength 1 */
  imagePath: string;
}

export interface OutfitPreview {
  items: ClothingItem[];
}

export interface SuggestOutfitBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  prompt: string;
}

export interface SuggestedOutfit {
  items: ClothingItem[];
  reasoning: string;
  suggestedName: string;
}

export interface SaveOutfitBody {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  itemIds: number[];
}

export interface Outfit {
  id: number;
  name: string;
  items: ClothingItem[];
  createdAt: string;
}

export type ClosetSummaryByCategoryItem = {
  category: Category;
  count: number;
};

export interface ClosetSummary {
  totalItems: number;
  savedOutfits: number;
  byCategory: ClosetSummaryByCategoryItem[];
}

export type ListItemsParams = {
  category?: Category;
};