# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Request a presigned URL for file upload
 */

export const RequestUploadUrlBody = zod.object({
  name: zod.string().min(1),
  size: zod.number().min(1),
  contentType: zod.string().min(1),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string().url(),
  objectPath: zod.string(),
  metadata: zod
    .object({
      name: zod.string().min(1),
      size: zod.number().min(1),
      contentType: zod.string().min(1),
    })
    .optional(),
});

/**
 * @summary Serve a public asset
 */
export const GetPublicObjectParams = zod.object({
  filePath: zod.coerce.string(),
});

/**
 * @summary Serve a private object
 */
export const GetStorageObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});

/**
 * @summary List all clothing items
 */
export const ListItemsQueryParams = zod.object({
  category: zod
    .enum(["top", "bottom", "shoes", "outerwear", "accessory"])
    .optional(),
});

export const ListItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
  color: zod.string().nullish(),
  imagePath: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListItemsResponse = zod.array(ListItemsResponseItem);

/**
 * @summary Create a new clothing item
 */

export const CreateItemBody = zod.object({
  name: zod.string().min(1),
  category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
  color: zod.string().nullish(),
  imagePath: zod.string().min(1),
});

export const CreateItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
  color: zod.string().nullish(),
  imagePath: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a clothing item
 */
export const DeleteItemParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteItemResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary List saved outfits
 */
export const ListOutfitsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
      color: zod.string().nullish(),
      imagePath: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  createdAt: zod.coerce.date(),
});
export const ListOutfitsResponse = zod.array(ListOutfitsResponseItem);

/**
 * @summary Save an outfit
 */

export const SaveOutfitBody = zod.object({
  name: zod.string().min(1),
  itemIds: zod.array(zod.number()).min(1),
});

export const SaveOutfitResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
      color: zod.string().nullish(),
      imagePath: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Generate a random outfit from the closet
 */
export const GenerateRandomOutfitResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
      color: zod.string().nullish(),
      imagePath: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary AI-powered outfit suggestion based on context
 */
export const suggestOutfitBodyPromptMax = 500;

export const SuggestOutfitBody = zod.object({
  prompt: zod.string().min(1).max(suggestOutfitBodyPromptMax),
});

export const SuggestOutfitResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
      color: zod.string().nullish(),
      imagePath: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  reasoning: zod.string(),
  suggestedName: zod.string(),
});

/**
 * @summary Delete a saved outfit
 */
export const DeleteOutfitParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteOutfitResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Closet summary counts by category
 */
export const GetClosetSummaryResponse = zod.object({
  totalItems: zod.number(),
  savedOutfits: zod.number(),
  byCategory: zod.array(
    zod.object({
      category: zod.enum(["top", "bottom", "shoes", "outerwear", "accessory"]),
      count: zod.number(),
    }),
  ),
});