# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Category = (typeof Category)[keyof typeof Category];

export const Category = {
  top: "top",
  bottom: "bottom",
  shoes: "shoes",
  outerwear: "outerwear",
  accessory: "accessory",
} as const;