# lib/api-zod/src/generated/types/closetSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClosetSummaryByCategoryItem } from "./closetSummaryByCategoryItem";

export interface ClosetSummary {
  totalItems: number;
  savedOutfits: number;
  byCategory: ClosetSummaryByCategoryItem[];
}