# lib/api-zod/src/generated/types/clothingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface ClothingItem {
  id: number;
  name: string;
  category: Category;
  color?: string | null;
  imagePath: string;
  createdAt: Date;
}