# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface CreateItemBody {
  /** @minLength 1 */
  name: string;
  category: Category;
  color?: string | null;
  /** @minLength 1 */
  imagePath: string;
}