# lib/api-zod/src/generated/types/outfit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItem } from "./clothingItem";

export interface Outfit {
  id: number;
  name: string;
  items: ClothingItem[];
  createdAt: Date;
}