# lib/api-zod/src/generated/types/outfitPreview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItem } from "./clothingItem";

export interface OutfitPreview {
  items: ClothingItem[];
}