# lib/api-zod/src/generated/types/saveOutfitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveOutfitBody {
  /** @minLength 1 */
  name: string;
  /** @minItems 1 */
  itemIds: number[];
}